fix: standardize V4 batch_id_per_token on int32 for fused SWA scatter
The fused decode SWA scatter loads batch_id_per_token at int32 width
(see ROCm/aiter#3793). The producers were int64, which raised
"batch_id_per_token must be 1-D int64" on the V4-Pro MTP decode path
(server failed to start -> accuracy job timed out).

Make all batch_id_per_token producers int32:
- v4_batch_id_per_token CpuGpuBuffer (model_runner path) int64 -> int32
- batch_id numpy sources (per-fwd + MTP draft) int64 -> int32
- sglang / vllm plugin bridge batch_id buffers + numpy sources -> int32

int32 indices are accepted by torch advanced-indexing (indexer meta) and
by the triton kernels (tl.load is dtype-agnostic); the explicit
.to(torch.int64) casts in csa_translate_pack / sglang remain and tolerate
int32 input. batch_id values are bounded by batch size, far below 2^31.

Validated end-to-end: DeepSeek-V4-Pro MTP3 GSM8K (3-shot) flexible
0.9477 / strict 0.9484, above the 0.94 CI threshold; decode drained
cleanly with no TypeError.

+++ atom/model_ops/v4_kernels/qk_norm_rope_maybe_quant.py
@@ -351,7 +351,7 @@ def qk_norm_rope_maybe_quant(
             BF16 only (requires ``quant_k=False``).
         state_slot_mapping: ``[bs]`` int32 — per-seq SWA ring slot. Required
             when ``swa_kv`` is set.
-        batch_id_per_token: ``[T]`` int64, ``-1`` on CG-pad tokens — token→seq
+        batch_id_per_token: ``[T]`` int32, ``-1`` on CG-pad tokens — token→seq
             map for the fused (flydsl) SWA scatter. Required by the flydsl path.
         swa_cu_seqlens_q: ``[bs+1]`` int — per-seq cumulative seqlens used by
             the Triton-fallback ``swa_write``. Required only on the fallback
